% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausbirthplace-package.R
\docType{package}
\name{ausbirthplace-package}
\alias{ausbirthplace}
\alias{ausbirthplace-package}
\title{ausbirthplace: Data for place of birth from the 2016 and 2019 Australian Census}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Michael \email{michael.lydeamore@monash.edu}

}
\keyword{internal}
