% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_birthplace}
\alias{census_birthplace}
\title{Number of Australian residents by country of birth from the 2016 and 2021 census}
\format{
\subsection{'census_birthplace'}{

A data frame with 105 rows and 4 columns
\describe{
\item{birth}{Country of birth of resident}
\item{count}{Number of residents from that birth country}
\item{percentage}{Percentage of residents from that birth country}
\item{census}{Year of census (either 2016 or 2021)}
}
}
}
\source{
\url{https://www.abs.gov.au/websitedbs/censushome.nsf/home/2016}
}
\usage{
census_birthplace
}
\description{
Number of Australian residents by country of birth from the 2016 and 2021 census
}
\keyword{datasets}
