% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number.R
\name{random_number}
\alias{random_number}
\title{Histograms of random numbers}
\usage{
random_number(range)
}
\arguments{
\item{range}{\strong{Maximum} value for the uniform distribution}
}
\value{
\code{ggplot2} histogram of the random values
}
\description{
Generates 100 uniform random numbers between 0 and \code{range}, and
plots a \code{ggplot2} histogram.
}
\details{
You can modify the plot using

\if{html}{\out{<div class="sourceCode r">}}\preformatted{random_number(5) + ggplot2::theme_bw()
}\if{html}{\out{</div>}}
}
\examples{
random_number(12)
}
